Script Python
I have a lot of trouble with this homework. Everything was running well til point 3. I had to ask my brother for some help to understand what's going on. On the internet, the whole info I found about was to use Pandas, which isn't supposed to be used on this module since we learned it til the next module.